optmized display trip
