+++ src/lib/components/ui/trips/display-trip/DisplayTrip.tsx
@@ -1,157 +1,24 @@
 'use client'
-import type { KindeUser } from '@kinde-oss/kinde-auth-nextjs'
-import { use } from 'react'
-import type { Trip } from '@/lib/types/types'
-import AddDefectForm from '../../add-defect/add-defect-form/AddDefectForm'
 import { BackLink } from '../../links/back-home-button'
 
 export default function DisplayTrip({
-	tripPromise,
-	userPromise,
+	tripTable,
+	pageHead,
+	addDefectForm,
 }: {
-	tripPromise: Promise<Trip>
-	userPromise: Promise<KindeUser<Record<string, string>> | null>
+	tripTable: React.ReactNode
+	pageHead: React.ReactNode
+	addDefectForm: React.ReactNode
 }) {
-	const trip = use(tripPromise)
-	const user = use(userPromise)
-
-	const twentyFourHoursAgo = new Date(Date.now() - 24 * 60 * 60 * 1000)
-
 	return (
 		<div className="flex w-full flex-col gap-4 rounded-sm bg-[var(--color-primary)] p-4 shadow-sm md:w-4/6">
 			<BackLink />
 
-			<PageHead
-				driverEmail={user?.email || ''}
-				driverName={user?.given_name || ''}
-			/>
+			{pageHead}
 
-			<table className="w-full overflow-hidden rounded-sm border p-4">
-				<TableHeadTrip />
-				<TableBody data={trip} />
-			</table>
+			{tripTable}
 
-			{trip?.date && new Date(trip.date) > twentyFourHoursAgo && (
-				<AddDefectForm
-					driverEmail={user?.email || ''}
-					tripId={Number(trip.tripid)}
-				/>
-			)}
+			{addDefectForm}
 		</div>
-	)
-}
-
-export function PageHead({
-	driverName,
-	driverEmail,
-}: {
-	driverName?: string
-	driverEmail?: string
-}) {
-	return (
-		<div className="w-full rounded-sm bg-[var(--color-background)] p-4 shadow-xl">
-			Driver Name: {driverName}
-			<br />
-			Driver Email: {driverEmail}
-		</div>
-	)
-}
-
-export function TableHeadTrip() {
-	return (
-		<thead>
-			<tr className="border-b">
-				<th className="w-1/3 rounded-sm p-2 text-left">Category</th>
-				<th className="w-1/3 rounded-sm p-2 text-left">Value</th>
-			</tr>
-		</thead>
-	)
-}
-
-function TableBody({ data }: { data: Trip }) {
-	return (
-		<tbody className="rounded-sm">
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Carrier</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.carrier}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					Carrier Address
-				</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.carrieraddress}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					Inspection Address
-				</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.inspectionaddress}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Truck Plate</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.truckplate}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					Trailer Plate A
-				</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.trailerplatea}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					Trailer Plate B
-				</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.trailerplateb}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Make</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">{data.make}</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Model</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">{data.model}</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Odometer</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.odometer}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Defects</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.defects}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">Remarks</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{data.remarks}
-				</td>
-			</tr>
-			<tr className="border-b">
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					Inspection Date
-				</td>
-				<td className="w-1/3 rounded-sm p-2 text-left">
-					{new Date(data.date).toLocaleString('en-CA', {
-						dateStyle: 'short',
-						timeStyle: 'short',
-					})}
-				</td>
-			</tr>
-		</tbody>
 	)
 }
